Connect result page to the store
