+++ frontend/src/pages/Result/container.js
@@ -4,8 +4,5 @@ import { connect } from 'react-redux';
 import View from './view';
 
 const mapStateToProps = (state) => ({ ...state.game });
-const actions = {
-  onCorrectAnswer: gameActions.handleCorrectAnswer,
-  onIncorrectAnswer: gameActions.handleIncorrectAnswer,
-};
+const actions = { onPlayAgainPress: gameActions.resetGame };
 export default connect(mapStateToProps, actions)(View);
